requirements.txt
unsloth
transformers
accelerate
peft
datasets
timm
pillow
faiss-gpu
torch
torchvision
tqdm
numpy
scikit-learn
pyyaml
